Added indent and outdent to the toolbar, some rearrangement for better composition
