+++ package.json
@@ -29,7 +29,7 @@
     "body-parser": "1.15.2",
     "bookshelf": "0.10.0",
     "draft-js": "^0.10.0",
-    "draftjs-utils": "^0.1.7",
+    "draftjs-utils": "^0.5.1",
     "express": "4.14.0",
     "jsonwebtoken": "7.1.9",
     "knex": "0.11.10",
